requirements.txt
python-telegram-bot==21.0.1
sentence-transformers==2.7.0
openai
numpy
python-dotenv==1.0.1
